+++ tasks/mpi/sadikov_I_sum_values_by_columns_matrix/src/mpi_src.cpp
@@ -20,8 +20,8 @@ bool sadikov_I_Sum_values_by_columns_matrix_mpi::MPITask::validation() {
 
 bool sadikov_I_Sum_values_by_columns_matrix_mpi::MPITask::pre_processing() {
   internal_order_test();
-  rows_count = taskData->inputs_count[0];
-  columns_count = taskData->inputs_count[1];
+  rows_count = static_cast<size_t>(taskData->inputs_count[0]);
+  columns_count = static_cast<size_t>(taskData->inputs_count[1]);
   matrix.reserve(rows_count * columns_count);
   auto *tmp_ptr = reinterpret_cast<int *>(taskData->inputs[0]);
   for (size_t i = 0; i < columns_count; ++i) {
@@ -67,8 +67,8 @@ bool sadikov_I_Sum_values_by_columns_matrix_mpi::MPITaskParallel::validation() {
 bool sadikov_I_Sum_values_by_columns_matrix_mpi::MPITaskParallel::pre_processing() {
   internal_order_test();
   if (world.rank() == 0) {
-    rows_count = taskData->inputs_count[0];
-    columns_count = taskData->inputs_count[1];
+    rows_count = static_cast<size_t>(taskData->inputs_count[0]);
+    columns_count = static_cast<size_t>(taskData->inputs_count[1]);
     delta = columns_count / world.size();
     last_column = columns_count % world.size();
     sum.reserve(columns_count);
@@ -131,8 +131,9 @@ bool sadikov_I_Sum_values_by_columns_matrix_mpi::MPITaskParallel::post_processin
   return true;
 }
 
-std::shared_ptr<ppc::core::TaskData> sadikov_I_Sum_values_by_columns_matrix_mpi::CreateTaskData(
-    std::vector<int> &InV, std::vector<size_t> &CeV, std::vector<int> &OtV) {
+std::shared_ptr<ppc::core::TaskData> sadikov_I_Sum_values_by_columns_matrix_mpi::CreateTaskData(std::vector<int> &InV,
+                                                                                                std::vector<int> &CeV,
+                                                                                                std::vector<int> &OtV) {
   auto taskData = std::make_shared<ppc::core::TaskData>();
   taskData->inputs.emplace_back(reinterpret_cast<uint8_t *>(InV.data()));
   taskData->inputs_count.emplace_back(CeV[0]);
