+++ tasks/seq/sadikov_I_sum_values_by_columns_matrix/func_tests/seq_func_tests.cpp
@@ -7,7 +7,7 @@
 
 TEST(sum_values_by_columns_matrix, check_validation1) {
   std::vector<int> in(144, 1);
-  std::vector<size_t> in_index{12, 12};
+  std::vector<int> in_index{12, 12};
   std::vector<int> out(12, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -17,7 +17,7 @@ TEST(sum_values_by_columns_matrix, check_validation1) {
 
 TEST(sum_values_by_columns_matrix, check_validation2) {
   std::vector<int> in(144, 1);
-  std::vector<size_t> in_index{12, 12};
+  std::vector<int> in_index{12, 12};
   std::vector<int> out(15, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -27,7 +27,7 @@ TEST(sum_values_by_columns_matrix, check_validation2) {
 
 TEST(sum_values_by_columns_matrix, check_empty_matrix) {
   std::vector<int> in(0);
-  std::vector<size_t> in_index{0, 0};
+  std::vector<int> in_index{0, 0};
   std::vector<int> out(0, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -43,7 +43,7 @@ TEST(sum_values_by_columns_matrix, check_empty_matrix) {
 
 TEST(sum_values_by_columns_matrix, check_square_matrix) {
   std::vector<int> in(144, 1);
-  std::vector<size_t> in_index{12, 12};
+  std::vector<int> in_index{12, 12};
   std::vector<int> out(12, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -59,7 +59,7 @@ TEST(sum_values_by_columns_matrix, check_square_matrix) {
 
 TEST(sum_values_by_columns_matrix, check_square_matrix2) {
   std::vector<int> in(256, 1);
-  std::vector<size_t> in_index{16, 16};
+  std::vector<int> in_index{16, 16};
   std::vector<int> out(16, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -75,7 +75,7 @@ TEST(sum_values_by_columns_matrix, check_square_matrix2) {
 
 TEST(sum_values_by_columns_matrix, check_square_matrix3) {
   std::vector<int> in(256, 1);
-  std::vector<size_t> in_index{16, 16};
+  std::vector<int> in_index{16, 16};
   std::vector<int> out(16, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -91,7 +91,7 @@ TEST(sum_values_by_columns_matrix, check_square_matrix3) {
 
 TEST(sum_values_by_columns_matrix, check_rect_matrix1) {
   std::vector<int> in(200, 1);
-  std::vector<size_t> in_index{10, 20};
+  std::vector<int> in_index{10, 20};
   std::vector<int> out(20, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -107,7 +107,7 @@ TEST(sum_values_by_columns_matrix, check_rect_matrix1) {
 
 TEST(sum_values_by_columns_matrix, check_rect_matrix2) {
   std::vector<int> in(500, 1);
-  std::vector<size_t> in_index{50, 10};
+  std::vector<int> in_index{50, 10};
   std::vector<int> out(10, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
@@ -123,7 +123,7 @@ TEST(sum_values_by_columns_matrix, check_rect_matrix2) {
 
 TEST(sum_values_by_columns_matrix, check_rect_matrix3) {
   std::vector<int> in(10000, 1);
-  std::vector<size_t> in_index{500, 20};
+  std::vector<int> in_index{500, 20};
   std::vector<int> out(20, 0);
   std::shared_ptr<ppc::core::TaskData> taskData =
       sadikov_I_Sum_values_by_columns_matrix_seq::CreateTaskData(in, in_index, out);
