+++ tasks/seq/sadikov_I_sum_values_by_columns_matrix/include/sq_task.h
@@ -10,7 +10,7 @@
 #include "core/task/include/task.hpp"
 // commit
 namespace sadikov_I_Sum_values_by_columns_matrix_seq {
-std::shared_ptr<ppc::core::TaskData> CreateTaskData(std::vector<int> &InV, const std::vector<size_t> &CeV,
+std::shared_ptr<ppc::core::TaskData> CreateTaskData(std::vector<int> &InV, const std::vector<int> &CeV,
                                                     std::vector<int> &OtV);
 std::vector<double> Randvector(size_t size);
 class MatrixTask : public ppc::core::Task {
